+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Switch hot EVM and mempool HashMaps to FxHashMap for faster hashing [#6303](https://github.com/lambdaclass/ethrex/pull/6303)
 
+### 2026-03-03
+
+- Replace RwLock/Mutex<HashMap> with DashMap for concurrent caches [#6297](https://github.com/lambdaclass/ethrex/pull/6297)
+
 ### 2026-03-02
 
 - SIMD-accelerate trie nibble operations for block execution [#6286](https://github.com/lambdaclass/ethrex/pull/6286)